Can't AI create my commit message?

+++ Starter-project/runtime-enabled-sdk/src/main/java/com/runtimeenabled/api/MyReSdkService.kt
@@ -30,5 +30,5 @@ interface MyReSdkService {
         callback: RemoteUiCallbackInterface
     ): SdkSandboxedUiAdapter
 
-    suspend fun createFile(sizeInMb: Int): String
+    suspend fun createFile(sizeInMb: Long): String
 }
